@@ -1,4 +1,5 @@
 <?php
+
 /*
  * This file is part of the "dragon-code/support" project.
  *
@@ -7,7 +8,7 @@
  *
  * @author Andrey Helldar <helldar@ai-rus.com>
  *
- * @copyright 2021 Andrey Helldar
+ * @copyright 2022 Andrey Helldar
  *
  * @license MIT
  *
@@ -18,7 +19,7 @@ namespace Tests\Fixtures\Instances;
 
 class Bar
 {
-    public $first = 'Foo';
+    public string $first = 'Foo';
 
-    public $second = 'Bar';
+    public string $second = 'Bar';
 }
