@@ -1,4 +1,5 @@
 <?php
+
 /*
  * This file is part of the "dragon-code/support" project.
  *
@@ -7,7 +8,7 @@
  *
  * @author Andrey Helldar <helldar@ai-rus.com>
  *
- * @copyright 2021 Andrey Helldar
+ * @copyright 2022 Andrey Helldar
  *
  * @license MIT
  *
@@ -20,9 +21,9 @@ use Tests\Fixtures\Contracts\Contract;
 
 class Bat implements Contract
 {
-    public $foo;
+    public ?string $foo;
 
-    public $bar;
+    public ?string $bar;
 
     public function __construct(?string $foo = null, ?string $bar = null)
     {
