@@ -7,7 +7,7 @@
  *
  * @author Andrey Helldar <helldar@ai-rus.com>
  *
- * @copyright 2021 Andrey Helldar
+ * @copyright 2022 Andrey Helldar
  *
  * @license MIT
  *
@@ -30,7 +30,7 @@ trait Castable
         }
     }
 
-    protected function castValue(string $key, $value)
+    protected function castValue(string $key, mixed $value)
     {
         $cast = $this->castKey($key);
 
@@ -41,7 +41,7 @@ trait Castable
 
     protected function castKey(string $key): string
     {
-        return $this->casts[$key] ?? 'string';
+        return $this->casts[$key] ?? 'default';
     }
 
     protected function castMethodName(string $key): string
@@ -49,7 +49,7 @@ trait Castable
         return (string) Str::of($key)->start('castTo_')->camel();
     }
 
-    protected function castToArray($value): array
+    protected function castToArray(mixed $value): array
     {
         if (empty($value)) {
             return [];
@@ -64,7 +64,7 @@ trait Castable
         return $output;
     }
 
-    protected function castToInteger($value): ?int
+    protected function castToInteger(mixed $value): ?int
     {
         return empty($value) && ! is_numeric($value) ? null : $value;
     }
@@ -72,5 +72,10 @@ trait Castable
     protected function castToString(?string $value): string
     {
         return (string) $value;
+    }
+
+    protected function castToDefault(mixed $value): mixed
+    {
+        return $value;
     }
 }
