@@ -7,7 +7,7 @@
  *
  * @author Andrey Helldar <helldar@ai-rus.com>
  *
- * @copyright 2021 Andrey Helldar
+ * @copyright 2022 Andrey Helldar
  *
  * @license MIT
  *
@@ -16,10 +16,12 @@
 
 namespace DragonCode\Support\Exceptions;
 
+use JetBrains\PhpStorm\Pure;
 use LogicException;
 
 class UnknownUrlComponentIndexException extends LogicException
 {
+    #[Pure]
     public function __construct(int $component)
     {
         parent::__construct('Unknown URL component index: ' . $component);
