@@ -7,7 +7,7 @@
  *
  * @author Andrey Helldar <helldar@ai-rus.com>
  *
- * @copyright 2021 Andrey Helldar
+ * @copyright 2022 Andrey Helldar
  *
  * @license MIT
  *
@@ -17,10 +17,8 @@
 namespace Tests;
 
 use DragonCode\Support\Facades\Facade;
-use DragonCode\Support\Facades\Helpers\Filesystem\Directory;
+use DragonCode\Support\Facades\Filesystem\Directory;
 use DragonCode\Support\Facades\Helpers\Str;
-use PHPUnit\Framework\Constraint\DirectoryExists;
-use PHPUnit\Framework\Constraint\LogicalNot;
 use PHPUnit\Framework\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase
@@ -37,13 +35,6 @@ abstract class TestCase extends BaseTestCase
         $this->destroyTempDirectory();
 
         parent::tearDown();
-    }
-
-    public static function assertDirectoryDoesNotExist(string $directory, string $message = ''): void
-    {
-        method_exists(get_parent_class(self::class), 'assertDirectoryDoesNotExist')
-            ? parent::assertDirectoryDoesNotExist($directory, $message)
-            : static::assertThat($directory, new LogicalNot(new DirectoryExists()), $message);
     }
 
     protected function tempDirectory(?string $path = null): string
@@ -68,9 +59,7 @@ abstract class TestCase extends BaseTestCase
     {
         $path = $this->tempDirectoryPrefix();
 
-        if (Directory::exists($path)) {
-            Directory::delete($path);
-        }
+        Directory::ensureDelete($path);
     }
 
     protected function tempDirectoryPrefix(): string
